DDF-5394 User settings is now extensible with extension points while keeping the plugin for downstream projects
DDF-5394 reset yarn.lock
