+++ ui/packages/catalog-ui-search/src/main/webapp/extension-points/index.tsx
@@ -17,3 +17,4 @@ export { default as routes } from './routes'
 export { default as navigator } from './navigator'
 export { default as searchInteractions } from './search-interactions'
 export { default as metacardInteractions } from './metacard-interactions'
+export { default as userSettings } from './user-settings'
